+++ src/data/products/products.json
@@ -178,5 +178,14 @@
     "category": "IPA",
     "size": "350 cc",
     "price": "593"
+  },
+  {
+    "id": "0a364fbd-c4d7-4199-9a92-c08c8ab59184",
+    "name": "Taylor",
+    "description": "Lala",
+    "category": "Champagne",
+    "size": "123",
+    "price": "123",
+    "image": "productPicture-1697474499472.jpg"
   }
 ]
